Remove the release.java.syncbase project
We had a discussion today and the decision was to not create a
separate git repo for the new Java API for Syncbase. Unlike JavaScript
and Go, we expect the users to use the jar from Maven/JCentral and not
build it from sources.

This is a rollback of http://v.io/c/22446.

Change-Id: I0e3223370eeb35ae62f3fea415ef9d1dee5b61ef